anihortes: Add PgUp / PgDn
PgUp and PgDn are mapped to Shift + Up and Shift + Down, i.e. drag
return on the Up and Down keys. This doesn't work very well because it
emits the Shift key along with PgUp and PgDn, and it makes it impossible
to use Up and Down with the Shift key (for marking text).

Change the behaviour, so that Shift + Up/Down now works for marking.

Instead map PgUp and PgDn to swiping north-east on i and south-east on s
respectively. This removes Enter and Space from those keys, but they
already have dedicated keys, so we don't need to have them twice.